Rename the airspace app to fxc-tiles
