+++ .eslintignore
@@ -1,2 +1,1 @@
 node_modules
-apps/airspaces/src/assets/airspaces
