+++ .prettierignore
@@ -4,6 +4,6 @@
 /coverage
 
 /apps/fxc-front/public
-/apps/airspaces/src/assets/airspaces
+/apps/fxc-tiles/src/assets/airspaces
 /.nx/cache
 /.nx/workspace-data
