--- apps/airspaces/jest.config.ts
@@ -1,17 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'airspaces',
-  preset: '../../jest.preset.js',
-  globals: {},
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]s$': [
-      'ts-jest',
-      {
-        tsconfig: '<rootDir>/tsconfig.spec.json',
-      },
-    ],
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/apps/airspaces',
-};
