--- apps/airspaces/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "airspaces",
-  "version": "1.0.0",
-  "description": "",
-  "scripts": {
-    "download": "chmod +x download.sh && ./download.sh"
-  },
-  "nx": {
-    "includedScripts": []
-  }
-}
