--- apps/airspaces/project.json
@@ -1,108 +1,0 @@
-{
-  "name": "airspaces",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "apps/airspaces/src",
-  "projectType": "application",
-  "tags": [],
-  "targets": {
-    "build": {
-      "executor": "@nx/webpack:webpack",
-      "outputs": ["{options.outputPath}"],
-      "dependsOn": [
-        {
-          "projects": ["secrets"],
-          "target": "build"
-        }
-      ],
-      "options": {
-        "target": "node",
-        "compiler": "tsc",
-        "outputPath": "dist/apps/airspaces",
-        "outputFileName": "unzip.js",
-        "main": "apps/airspaces/src/app/unzip.ts",
-        "generatePackageJson": true,
-        "additionalEntryPoints": [
-          {
-            "entryName": "create-geojson",
-            "entryPath": "apps/airspaces/src/app/create-geojson.ts"
-          },
-          {
-            "entryName": "create-tiles",
-            "entryPath": "apps/airspaces/src/app/create-tiles.ts"
-          },
-          {
-            "entryName": "download-openaip",
-            "entryPath": "apps/airspaces/src/app/download-openaip.ts"
-          },
-          {
-            "entryName": "create-tiles-info",
-            "entryPath": "apps/airspaces/src/app/create-tiles-info.ts"
-          },
-          {
-            "entryName": "stats",
-            "entryPath": "apps/airspaces/src/app/stats.ts"
-          },
-          {
-            "entryName": "create-tiles-info-diff",
-            "entryPath": "apps/airspaces/src/app/create-tiles-info-diff.ts"
-          },
-          {
-            "entryName": "upload-tiles-diff",
-            "entryPath": "apps/airspaces/src/app/upload-tiles-diff.ts"
-          }
-        ],
-        "tsConfig": "apps/airspaces/tsconfig.app.json",
-        "assets": [
-          {
-            "glob": "download.sh",
-            "input": "apps/airspaces/src/assets/",
-            "output": ""
-          },
-          {
-            "glob": "cors.json",
-            "input": "apps/airspaces/src/assets/",
-            "output": "assets"
-          }
-        ],
-        "babelUpwardRootMode": true,
-        "webpackConfig": "apps/airspaces/webpack.config.js"
-      },
-      "configurations": {
-        "production": {
-          "optimization": true,
-          "extractLicenses": true,
-          "inspect": false
-        }
-      }
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "dependsOn": [
-        {
-          "projects": ["secrets"],
-          "target": "build"
-        }
-      ],
-      "options": {
-        "jestConfig": "apps/airspaces/jest.config.ts"
-      }
-    },
-    "container": {
-      "executor": "@nx-tools/nx-container:build",
-      "dependsOn": ["build"],
-      "options": {
-        "engine": "docker",
-        "pull": true,
-        "push": true,
-        "tags": ["gcr.io/fly-xc/unzip"]
-      }
-    },
-    "serve": {
-      "executor": "nx:run-commands",
-      "options": {
-        "command": "npx servez -p 8084 -q apps/airspaces/src/assets/airspaces/tiles"
-      }
-    }
-  }
-}
