--- apps/airspaces/src/app/util.ts
@@ -1,5 +1,0 @@
-export function printOnCurrentLine(msg: string) {
-  process.stdout.clearLine(0);
-  process.stdout.cursorTo(0);
-  process.stdout.write(msg);
-}
