--- apps/airspaces/src/assets/airspaces/.gitignore
@@ -1,6 +1,0 @@
-/openaip.json
-/UKRAINE (UK).txt
-/airspaces.geojson
-/tiles
-/tiles-info-previous.json
-/tiles-info-diff.json
