--- apps/airspaces/src/assets/cors.json
@@ -1,18 +1,0 @@
-[
-  {
-    "origin": [
-      "http://localhost:8080",
-      "https://console.cloud.google.com",
-      "https://fly-xc.appspot.com/",
-      "https://flyxc.app/"
-    ],
-    "responseHeader": [
-      "Content-Type",
-      "Content-Encoding"
-    ],
-    "method": [
-      "GET"
-    ],
-    "maxAgeSeconds": 360000
-  }
-]
