--- apps/airspaces/webpack.config.js
@@ -1,8 +1,0 @@
-const { composePlugins, withNx } = require('@nx/webpack');
-
-// Nx plugins for webpack.
-module.exports = composePlugins(withNx(), (config) => {
-  // Note: This was added by an Nx migration. Webpack builds are required to have a corresponding Webpack config file.
-  // See: https://nx.dev/recipes/webpack/webpack-config-setup
-  return config;
-});
