+++ apps/fxc-tiles/project.json
@@ -1,0 +1,108 @@
+{
+  "name": "fxc-tiles",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "apps/fxc-tiles/src",
+  "projectType": "application",
+  "tags": [],
+  "targets": {
+    "build": {
+      "executor": "@nx/webpack:webpack",
+      "outputs": ["{options.outputPath}"],
+      "dependsOn": [
+        {
+          "projects": ["secrets"],
+          "target": "build"
+        }
+      ],
+      "options": {
+        "target": "node",
+        "compiler": "tsc",
+        "outputPath": "dist/apps/fxc-tiles",
+        "outputFileName": "unzip.js",
+        "main": "apps/fxc-tiles/src/app/airspaces/unzip.ts",
+        "generatePackageJson": true,
+        "additionalEntryPoints": [
+          {
+            "entryName": "airspaces/create-geojson",
+            "entryPath": "apps/fxc-tiles/src/app/airspaces/create-geojson.ts"
+          },
+          {
+            "entryName": "airspaces/create-tiles",
+            "entryPath": "apps/fxc-tiles/src/app/airspaces/create-tiles.ts"
+          },
+          {
+            "entryName": "airspaces/download-openaip",
+            "entryPath": "apps/fxc-tiles/src/app/airspaces/download-openaip.ts"
+          },
+          {
+            "entryName": "airspaces/create-tiles-info",
+            "entryPath": "apps/fxc-tiles/src/app/airspaces/create-tiles-info.ts"
+          },
+          {
+            "entryName": "airspaces/stats",
+            "entryPath": "apps/fxc-tiles/src/app/airspaces/stats.ts"
+          },
+          {
+            "entryName": "airspaces/create-tiles-info-diff",
+            "entryPath": "apps/fxc-tiles/src/app/airspaces/create-tiles-info-diff.ts"
+          },
+          {
+            "entryName": "airspaces/upload-tiles-diff",
+            "entryPath": "apps/fxc-tiles/src/app/airspaces/upload-tiles-diff.ts"
+          }
+        ],
+        "tsConfig": "apps/fxc-tiles/tsconfig.app.json",
+        "assets": [
+          {
+            "glob": "download.sh",
+            "input": "apps/fxc-tiles/src/assets/airspaces",
+            "output": "airspaces"
+          },
+          {
+            "glob": "cors.json",
+            "input": "apps/fxc-tiles/src/assets/airspaces",
+            "output": "assets"
+          }
+        ],
+        "babelUpwardRootMode": true,
+        "webpackConfig": "apps/fxc-tiles/webpack.config.js"
+      },
+      "configurations": {
+        "production": {
+          "optimization": true,
+          "extractLicenses": true,
+          "inspect": false
+        }
+      }
+    },
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "dependsOn": [
+        {
+          "projects": ["secrets"],
+          "target": "build"
+        }
+      ],
+      "options": {
+        "jestConfig": "apps/fxc-tiles/jest.config.ts"
+      }
+    },
+    "container": {
+      "executor": "@nx-tools/nx-container:build",
+      "dependsOn": ["build"],
+      "options": {
+        "engine": "docker",
+        "pull": true,
+        "push": true,
+        "tags": ["gcr.io/fly-xc/unzip"]
+      }
+    },
+    "serve": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "npx servez -p 8084 -q apps/fxc-tiles/src/assets/airspaces/tiles"
+      }
+    }
+  }
+}
