+++ apps/fxc-tiles/src/assets/airspaces/.gitignore
@@ -1,0 +1,6 @@
+/openaip.json
+/UKRAINE (UK).txt
+/airspaces.geojson
+/tiles
+/tiles-info-previous.json
+/tiles-info-diff.json
